Replace all Windows with Pages.
If I understand correctly (which I may not) we only want 1 Window class and the rest should be "Pages" that we navigate between using the <hyperlink> element, or a "navigation service". To that end I have replaced all "Windows" with "Pages" and just left the MainWindow as a shell that contains the pages. The TabBar has been updated to use some of these <Hyperlink> elements to demonstrate them.