Add a set of model classes to the project
This commit creates a set of model classes that can represent the data for the application in memory (we don't have to persist this data anywhere).

The main entry point for this data is the User class. This class should be accessed by calling User.getInstance() to create a singleton instance (this means only one instance of the class ever exists). The rest of the data can be accessed by accessing the various properties of on the User class (Meals, Steps, Actities, Info, Goal). The beginning of some fake "historical data" methods have been written, currently they just create some meal entries but we should add more as we need them.

I have started adopting the "active record" pattern on the Model.Meal class where instance methods/properties refer to a single instance of a meal and the static methods are helper methods to help you get Meal objects, for example "Meal.getMealsByDate" gets all meals matching the current date. We should continue this pattern whenever we want to find items, for example if someone wanted to get all activities for the past month they should add the relevant static method to the Activity class.